feat: 파일 view/controller/domain 으로 분리
