+++ src/view/icon.js
