+++ src/view/select.js
@@ -1,5 +1,5 @@
 import { render } from "./render.js";
-import { sortingHandler, filteringHandler } from "./selectAction.js";
+import { sortingHandler, filteringHandler } from "../controller/cardController.js";
 
 export const addSelectEvent = () => {
   const sort = document.querySelector("#sorting-filter");
